requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
autobahn==18.11.2
Automat==0.7.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
coinmarketcap-api==0.0.2
beautifulsoup4==4.7.1
bs4==0.0.1
requests-cache==0.5.0
matplotlib==3.1.0
cryptography==2.4.2
dateparser==0.7.0
dbf-light==0.3.0
Events==0.3
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycbrf==0.3.0
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pyTelegramBotAPI==3.6.6
python-binance==0.7.0
python-binance-api==1.1.0
python-bittrex==0.3.0
python-dateutil==2.7.5
python-kucoin==0.1.12
pytz==2018.7
regex==2018.11.22
requests==2.20.1
service-identity==17.0.0
simplejson==3.16.0
six==1.11.0
Twisted==18.9.0
txaio==18.8.1
tzlocal==1.5.1
urllib3==1.24.1
websocket-client==0.54.0
zope.interface==4.6.0
